Agregado cartel de error al inscribirse a examen

+++ alumno/inscripcion/js/inscripcionExamen.js
@@ -34,11 +34,19 @@ function start(auth){
 
     initTable();
     table.setOnDeleteEvent((id) => {
-        axios.delete(api.examen.inscripcion+"/"+id, getHeader())
-        .then((response) => {if(response.status == 200){
-                searchApi();
-            }}
-        );
+        axios.get(api.examen.examen + "/" + id, getHeader()
+        ).then(function (response) {
+            if(response.status == 200){
+                    if (response.data.inscripcion=="SI"){
+                    axios.delete(api.examen.inscripcion+"/"+id, getHeader())
+                    .then((response) => {if(response.status == 200){
+                            searchApi();
+                        }}
+                    );
+                    } else {
+                        alert("Error: El usuario no esta inscripto al examen");
+                    }
+            }
     });
 
     table.setOnEditEvent((id) =>{
